requirements/base.txt
Django==3.2.7
psycopg2==2.9.1
djangorestframework==3.12.4
requests==2.26.0
